Moving selectors into common folder
