+++ src/applications/personalization/common/selectors/index.js
