+++ src/applications/personalization/common/selectors/ratedDisabilities.js
