+++ src/applications/personalization/dashboard/components/Dashboard.jsx
@@ -39,7 +39,7 @@ import MPIConnectionError from '~/applications/personalization/components/MPICon
 import NotInMPIError from '~/applications/personalization/components/NotInMPIError';
 import IdentityNotVerified from '~/applications/personalization/components/IdentityNotVerified';
 import { fetchTotalDisabilityRating as fetchTotalDisabilityRatingAction } from '~/applications/personalization/profile/actions/ratedDisabilities';
-import { hasTotalDisabilityServerError } from '../selectors/ratedDisabilities';
+import { hasTotalDisabilityServerError } from '../../common/selectors/ratedDisabilities';
 import { API_NAMES } from '../../common/constants';
 import useDowntimeApproachingRenderMethod from '../useDowntimeApproachingRenderMethod';
 import ApplyForBenefits from './apply-for-benefits/ApplyForBenefits';
