--- src/applications/personalization/profile/selectors/ratedDisabilities.js
@@ -1,13 +1,0 @@
-import { isServerError } from '../util';
-
-export const totalDisabilityError = state => {
-  return state.totalRating?.error ?? null;
-};
-
-export const hasTotalDisabilityServerError = state => {
-  const error = totalDisabilityError(state);
-  if (!error) {
-    return false;
-  }
-  return isServerError(error.code);
-};
